src/mcp_server/pyproject.toml
[project]
name = "mcp-docs-server"
version = "0.1.0"
description = "MCP server for documentation search"
dependencies = [
    "fastmcp",
    "httpx",
    "beautifulsoup4",
    "python-dotenv",
    "loguru>=0.7.3",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
